mac80211: generate hw flags from include file
Instead of having to list all the hw flags in many places, just
list the macro to generate everything in a new include file and
include it in the proper places.

For now this just replaces them in a single place in debugfs,
but the next patch will add more places.

I've not replaced the actual flags definitions since docbook
will otherwise not be happy. Forgetting to update one or the
other is an instant compiler failure though (after the next
patch) so that's not really problematic.

Signed-off-by: Johannes Berg <johannes.berg@intel.com>